message triggers #4272 #4273
